backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
pandas>=1.5
reportlab>=4.0
